pyproject.toml
[project]
name = "hetriever"
version = "0.1.0"
description = "Hugo documentation retriever with vector embeddings"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb",
    "click>=8.3.0",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-benchmark",
    "hypothesis",
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "DTZ", "T20", "RET", "SIM", "ARG", "PTH", "PL", "RUF"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["PLR2004", "S101", "ARG"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"
markers = [
    "contract: contract tests",
    "integration: integration tests",
    "unit: unit tests",
    "benchmark: performance benchmarks",
    "property: property-based tests",
    "edge_case: edge case tests",
]
